+++ xbmc/pvr/dialogs/GUIDialogPVRChannelsOSD.cpp
@@ -266,7 +266,7 @@ void CGUIDialogPVRChannelsOSD::Clear()
 
 void CGUIDialogPVRChannelsOSD::GotoChannel(int item)
 {
-  if (item < 0 || item >= (int)m_vecItems->Size()) {
+  if (item < 0 || item >= m_vecItems->Size()) {
     return;
 }
 
@@ -277,7 +277,7 @@ void CGUIDialogPVRChannelsOSD::GotoChannel(int item)
 
 void CGUIDialogPVRChannelsOSD::ShowInfo(int item)
 {
-  if (item < 0 || item >= (int)m_vecItems->Size()) {
+  if (item < 0 || item >= m_vecItems->Size()) {
     return;
 }
